add(modules): divide structure into modules
